fixed standalone component issue

+++ src/app/app.ts
@@ -3,6 +3,7 @@ import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'app-root',
+  standalone: true,
   imports: [RouterOutlet],
   templateUrl: './app.html',
   styleUrl: './app.css'
@@ -11,7 +12,10 @@ export class App {
   protected readonly title = signal('user-management-system');
 
   logout() {
-    localStorage.clear();
-    window.location.href = '/login';
+    if (typeof window !== 'undefined') {
+      localStorage.clear();
+      window.location.href = '/login';
+    }
   }
+
 }
